Add Meta Viewport Header and Ported OL and UL IDs
Add correct viewport header to HTML files to enable correct viewport in browser. HTML standards
prefer the viewport to be defined in the header using the Meta tag. Also ported the OL and UL
IDs to classes. These IDs were initially coded as IDs for a learning exercise. Considering the IDs
were intended to provide site-wide styling of the ordered and unordered lists (specifically the ingredients and recipe steps)
they should be classes so the code can be re-sued as needed.