Adjust padding and margins on inline elements
Adjusting the padding and margins on inline elements allows flex to work properly and maintain a clean appearance.
The style sheet is still on reflected on the Lasagna page, the image-flex class still needs added to the img tags on the other pages.
There is still undesirable behavior with elements that have flex added to them. Both OL and UL classes have lost bullet points or numbers, breaking the list. Will look into this behavior in a future update.